% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.tbl_kgx.R
\name{ancestors.tbl_kgx}
\alias{ancestors.tbl_kgx}
\title{Expand ancestors transitively}
\usage{
\method{ancestors}{tbl_kgx}(g, ...)
}
\arguments{
\item{g}{A \code{tbl_kgx()} graph to expand.}

\item{...}{Other parameters (unused).}
}
\value{
A tbl_kgx graph.
}
\description{
Expand a tbl_kgx graph to include all ancestors of nodes defined transitively
by "biolink:subclass_of" relationships. This is a simple wrapper around
expand(predicates = "biolink:subclass_of", direction = "out", transitive = TRUE)
}
\examples{
engine <- file_engine(system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr"))
g <- engine |>
	 fetch_nodes(query_ids = "MONDO:0020066") |>
	 ancestors()

}
