% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.monarch_kg.R
\name{as_monarch_kg}
\alias{as_monarch_kg}
\title{Convert a tidygraph to a monarch_kg graph.}
\usage{
as_monarch_kg(graph_obj, ...)
}
\arguments{
\item{graph_obj}{input to convert to a monarch_kg.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a new monarch_kg.
}
\description{
Converts a tidygraph of the appropriate format into a monarch_kg graph. Not intended for direct use.
}
\details{
The monarch_kg class extends tidygraph and defines Monarch-specific methods in addition to functionality provided by tidygraph and igraph.
}
\examples{
g <- monarch_search("Cystic Fibrosis", limit = 3)

str(g)

print(methods(class = "monarch_kg"))
}
\seealso{
\code{\link{monarch_search}}.
}
