% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_engine.R
\name{base_engine}
\alias{base_engine}
\title{base_engine}
\usage{
base_engine(name = "default_engine", preferences = NULL, ...)
}
\arguments{
\item{name}{A character string indicating the name of the engine.}

\item{preferences}{A named list of preferences for the engine.}
}
\value{
An object of class \code{base_engine}
}
\description{
A base class for all engines
}
\examples{
base_engine()
base_engine(name = "test")
base_engine(preferences = "preferences.yaml")
}
