% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_disease_assoc_w_gene}
\alias{bioentity_disease_assoc_w_gene}
\title{Diseases associated with a gene.}
\usage{
bioentity_disease_assoc_w_gene(gene)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of homolog information, monarch_api S3 class).
}
\description{
Given a gene, what diseases are associated with it.
}
\details{
https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene%3A8314/diseases/?rows=100&fetch_objects=true&format=json
}
\examples{
gene <- "NCBIGene:8314"
bioentity_disease_assoc_w_gene(gene)
}
