% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_gene_homology_associations}
\alias{bioentity_gene_homology_associations}
\title{Gets homolog associations for a gene.}
\usage{
bioentity_gene_homology_associations(gene)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of homolog information, monarch_api S3 class).
}
\description{
The monarch_api class is included in return mainly for debugging.
}
\details{
https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene%3A8314?rows=100&fetch_objects=true
}
\examples{
gene <-"NCBIGene:8314"
bioentity_gene_homology_associations(gene)
}
