% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_gene_info}
\alias{bioentity_gene_info}
\title{Gets gene info, like catgories and labels for a gene.}
\usage{
bioentity_gene_info(gene, rows = 100)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (list of gene information, monarch_api S3 class).
}
\description{
The monarch_api class is included in return mainly for debugging the REST requests.
}
\details{
https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene%3A8314?rows=100&fetch_objects=true
https://api.monarchinitiative.org/api/bioentity/gene/HGNC%3A950?fetch_objects=true&rows=100
}
\examples{
gene <-"NCBIGene:8314"
bioentity_gene_info(gene)
}
