% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_homologs}
\alias{bioentity_homologs}
\title{Gets homologs to a gene.}
\usage{
bioentity_homologs(gene)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of homolog information, monarch_api S3 class).
}
\description{
The monarch_api class is included in return mainly for debugging.
}
\examples{
gene <-"NCBIGene:8314"
bioentity_homologs(gene)
}
