% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_id_type}
\alias{bioentity_id_type}
\title{Gets info of a 'type' from an 'id'}
\usage{
bioentity_id_type(id, type, rows = 100)
}
\arguments{
\item{id}{A valid monarch initiative id.}

\item{type}{A category to try for the id.}
}
\value{
A list of (list of info for the id, monarch_api S3 class).
}
\description{
WARNING: will return a result even if that result doesn't seem to match the type.
}
\details{
This is a general method for when other appropriate API calls may not be available.
Or you don't want to bother discovering the appropriate function name.

As a side effect of our Warning, in the event of an unsuccessful mapping, the info
returned  might still be useful for discovery of what categories your id maps to.

mimics https://api.monarchinitiative.org/api/bioentity/disease/MONDO%3A0006486?fetch_objects=true&rows=100

Types seem to map to Monarch Initiative 'categories' and are limited(?) to:

- gene
    - variant
    - disease
    - genotype
    - phenotype
    - goterm
    - pathway
    - anatomy
    - substance
    - individual
}
\examples{
disease <-"MONDO:0006486"
bioentity_id_type(disease, 'disease')
#' # Note in this example, the info returned includes 'categories': 'disease' and 'quality'
# This is a hint of which valid 'types' will be 'successful' for a particular 'id'.
# Warning: Trying other types returns same info, because the other categories
# were irrelevant for this id.
# bioentity_id_type(disease, 'gene') # Not run. Does not get 'gene'!
}
