% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_interactions_assoc_w_gene}
\alias{bioentity_interactions_assoc_w_gene}
\title{Gets interactions associated with a gene.}
\usage{
bioentity_interactions_assoc_w_gene(gene)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of variant information, monarch_api S3 class).
}
\description{
Given a gene, what variants are associated with it.
}
\details{
https://api.monarchinitiative.org/api/bioentity/gene/HGNC%3A950/interactions/?rows=100&fetch_objects=true
}
\examples{
gene <- "NCBIGene:8314"
bioentity_interactions_assoc_w_gene(gene)
}
