% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_phenotypes_assoc_w_gene}
\alias{bioentity_phenotypes_assoc_w_gene}
\title{Gets phenotypes associated with a gene.}
\usage{
bioentity_phenotypes_assoc_w_gene(gene, rows = 100)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of phenotype information, monarch_api S3 class).
}
\description{
Given a gene, what phenotypes are associated with it.
}
\details{
https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene%3A8314/phenotypes/?rows=100&fetch_objects=true
}
\examples{
gene <- "NCBIGene:8314"
bioentity_phenotypes_assoc_w_gene(gene)
}
