% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{biolink_search}
\alias{biolink_search}
\title{Get a list of results matching a search term.}
\usage{
biolink_search(phrase_or_id, rows = 20)
}
\arguments{
\item{phrase_or_id}{Something for which we are looking for more information, or to find useful entities.}
}
\value{
A list of (the search results, and monarch_api S3 class.)
}
\description{
https://api.monarchinitiative.org/api/search/entity/uveal%20melanoma?rows=20&start=1&category=disease
}
\details{
Note, as opposed to the bioentity_ functions, the biolink_search results are not standardized to a tibble?

https://api.monarchinitiative.org/api/search/entity/autocomplete/MELANOMA%2C%20UVEAL?rows=20&start=1&category=disease # finds MONDO:0006486 as 3rd result.
}
\examples{
term <- "uveal melanoma"
biolink_search(term)
}
