% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_monarch_url}
\alias{build_monarch_url}
\title{Builds URL for a monarch GET request.}
\usage{
build_monarch_url(path, query = NULL)
}
\arguments{
\item{path}{A path as a list to the monarch resource to use.}

\item{query}{A list of url parameter settings. TRUE FALSE set to
"true" "false".}
}
\value{
URL as a safely encoded string.
}
\description{
The path needs to be safely encoded.
}
\details{
But the query apparently will be re-encoded, even if it is already encoded.
So it is best not to do any or our own encoding on the query list.

url <- httr::modify_url("https://api.monarchinitiative.org",
           path = "test%3Ame",
           query = list(x='this_is:ok', y="but_this_is%3Anotok"))
    "https://api.monarchinitiative.org/test%3Ame?x=this_is%3Aok&y=but_this_is%253Anotok"
}
\examples{
m_path <- "/api/bioentity/gene"
gene <- "NCBIGene\%3A8314"
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = "true",
                         format = "json"))
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = TRUE,
                         unselect_evidence=FALSE,
                         format = "json"))
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = TRUE,
                         unselect_evidence=NULL,
                         format = "json"))
}
\seealso{
\link[utils]{URLencode}
}
