% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_query}
\alias{clean_query}
\title{Converts R booleans to monarch_api values and drops NULL parameters from query list.}
\usage{
clean_query(query)
}
\arguments{
\item{query}{List of api query parameters and their values.}
}
\value{
list of query paramters and cleaned values.
}
\description{
Converts R booleans to monarch_api values and drops NULL parameters from query list.
}
\examples{
clean_query(list(rows = 100,
                 fetch_objects = TRUE,
                 use_compact_associations = FALSE,
                 unselect_evidence=NULL, format = "json"))
}
