% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query_df.R
\name{cypher_query_df}
\alias{cypher_query_df}
\title{Execute a Cypher Query}
\usage{
cypher_query_df(engine, query, parameters = NULL, ...)
}
\arguments{
\item{engine}{A neo4j_engine() or derivative providing access to a Neo4j
database.}

\item{query}{A string representing the Cypher query, which should return a
table. Multiple queries may be passed as a vector; if so,
Neo2R::multicypher if used and the result is returned as a list
of data frames.}

\item{parameters}{A list of parameters for the Cypher query, if required.}

\item{...}{Additional arguments passed to the function.}
}
\value{
The result of the Cypher query as a data frame, or a list of data
frames if multiple queries are passed.
}
\description{
This function takes a Cypher query and parameters, executes the query using
the given engine, and returns the result as a data frame.
}
\examples{
\dontshow{if (monarch_engine_check()) withAutoprint(\{ # examplesIf}
engine <- monarch_engine()

query <- "MATCH (n) WHERE n.id IN $ids RETURN n LIMIT 10"
parameters <- list(ids = c("MONDO:0007525", 
                           "MONDO:0020066", 
                           "MONDO:0034021"))

result <- cypher_query_df(engine, query, parameters)
print(result)
\dontshow{\}) # examplesIf}
}
