% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{cytoscape}
\alias{cytoscape}
\title{Send a graph to Cytoscape}
\usage{
cytoscape(g, ...)
}
\arguments{
\item{g}{A \code{tbl_kgx()} graph to visualize.}

\item{...}{Other parameters (unused).}
}
\value{
NULL, invisibly
}
\description{
Given a tbl_kgx graph, send it to Cytoscape for visualization. Node labels
are mapped to node \code{name} (if available, otherwise they default to node
\code{id}),
node color is mapped to \code{pcategory}, edge color is mapped to \code{predicate},
node hover-over text is set to \code{description} (if available, otherwise node
\code{id}),
and edge hover-over text is set to \code{predicate}. Nodes are layed out
using the Kamada-Kawai method. These properties and more may be customized in
the Cytoscape application. This function requires that Cytoscape is installed
and running independently of the R session.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
data(eds_marfan_kg)
g <- eds_marfan_kg |>
    fetch_nodes(query_ids = "MONDO:0020066") |>
    expand(predicates = "biolink:subclass_of", 
           direction = "in", 
           transitive = TRUE) |>
    expand(categories = c("biolink:PhenotypicFeature", "biolink:Gene"))

# Cytoscape must be installed and running
cytoscape(g)
\dontshow{\}) # examplesIf}
}
