% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendants}
\alias{descendants}
\title{Expand descendants transitively}
\usage{
descendants(g, ...)
}
\arguments{
\item{g}{A \code{tbl_kgx()} graph to expand.}

\item{...}{Other parameters (unused).}
}
\value{
A tbl_kgx graph.
}
\description{
Expand a tbl_kgx graph to include all descendants of nodes defined transitively
by "biolink:subclass_of" relationships. This is a simple wrapper around
expand(predicates = "biolink:subclass_of", direction = "in", transitive = TRUE)
}
\examples{
data(eds_marfan_kg)
g <- eds_marfan_kg |>
	 fetch_nodes(query_ids = "MONDO:0020066") |>
	 descendants()

}
