% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edges.tbl_kgx}
\alias{edges.tbl_kgx}
\title{Get tbl_kgx graph edges table.}
\usage{
\method{edges}{tbl_kgx}(x, ...)
}
\arguments{
\item{x}{Input graph}

\item{...}{Other options (unused)}
}
\description{
Get tbl_kgx graph edges table.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 1)
print(edges(g))
}
