% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graph.file_engine.R
\name{example_graph.file_engine}
\alias{example_graph.file_engine}
\title{Return an example set of nodes from a KG engine.}
\usage{
\method{example_graph}{file_engine}(engine, ...)
}
\arguments{
\item{engine}{A \code{file_engine} object}

\item{...}{Other parameters (not used)}
}
\value{
A tbl_kgx graph
}
\description{
Given a KGX file-based KG engine, returns a graph representing the diversity
of node categories and edge predicates for browsing. The returned graph is guaranteed to
contain at least one node of every category, and at least one edge of every
predicate. No other guarantees are made: the example graph is not minimal
to satisfy these criteria, it is not random or even pseudo-random, and it
may not be connected.
}
\examples{
# Using example KGX file packaged with monarchr
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")

# Retrieve and print an example graph:
g <- file_engine(filename) |> example_graph()
print(g)
}
