% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graph.neo4j_engine.R
\name{example_graph.neo4j_engine}
\alias{example_graph.neo4j_engine}
\title{Return an example set of nodes from a KG engine.}
\usage{
\method{example_graph}{neo4j_engine}(engine, ...)
}
\arguments{
\item{engine}{A \code{neo4j_engine} object}

\item{...}{Other parameters (not used)}
}
\value{
A tbl_kgx graph
}
\description{
Given a KGX Neo4j KG engine, returns a graph representing the diversity
of node categories and edge predicates for browsing. The returned graph is
guaranteed to
contain at least one node of every category, and at least one edge of every
predicate. No other guarantees are made: the example graph is not minimal
to satisfy these criteria, it is not random or even pseudo-random, and it
may not be connected.
}
\examples{
\dontshow{if (monarch_engine_check()) withAutoprint(\{ # examplesIf}
# Retrieve and print an example graph:
g <- monarch_engine() |> example_graph()
print(g)
\dontshow{\}) # examplesIf}
}
