% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Fetch additional knowledge graph edges connected to a query graph}
\usage{
expand(
  graph,
  engine = NULL,
  direction = "both",
  predicates = NULL,
  categories = NULL,
  transitive = FALSE,
  ...
)
}
\arguments{
\item{graph}{A query \code{tbl_kgx()} graph ot query with.}

\item{engine}{(Optional) An engine to use for fetching query graph edges. If
not provided, the graph's most recent engine is used.}

\item{direction}{The direction of associations to fetch. Can be "in", "out",
or "both". Default is "both".}

\item{predicates}{A vector of relationship predicates (nodes in g are
subjects in the KG), indicating which edges to consider in
the neighborhood. If NULL (default), all edges are
considered.}

\item{categories}{A vector of node categories, indicating which nodes in the
larger KG may be fetched. If NULL (default), all nodes in
the larger KG are will be fetched.}

\item{transitive}{If TRUE, include transitive closure of the neighborhood.
Default is FALSE. Useful in combination with predicates
like \code{biolink:subclass_of}.}

\item{...}{Other parameters passed to methods.}
}
\value{
A \code{tbl_kgx()} graph
}
\description{
Given an optional KG engine (e.g. a \code{file_engine()},
\code{neo4j_engine()}, or \code{monarch_engine()}) and a query \code{tbl_kgx()} graph,
fetches additional nodes and edges
from the KG, expanding the query graph according to specific criteria. If
the first parameter is an engine, that
engine is used; if the first parameter is a query graph, the most recent
engine associated with the graph is used.
}
\examples{
\dontshow{if (monarch_engine_check()) withAutoprint(\{ # examplesIf}
## Using Monarch (hosted)
phenos <- monarch_engine() |>
    fetch_nodes(query_ids = "MONDO:0007525") |>
    expand(
        predicates = "biolink:has_phenotype",
        categories = "biolink:PhenotypicFeature"
    )

print(phenos)
\dontshow{\}) # examplesIf}
## Using example KGX file packaged with monarchr
data(eds_marfan_kg)
phenos <- eds_marfan_kg |>
    fetch_nodes(query_ids = "MONDO:0007525") |>
    expand(
        predicates = "biolink:has_phenotype",
        categories = "biolink:PhenotypicFeature"
    )

print(phenos)
}
