% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explode}
\alias{explode}
\title{Explode a graph into a list of single-node graphs.}
\usage{
explode(x, ...)
}
\arguments{
\item{x}{Input graph}

\item{...}{Other options (unused)}
}
\value{
A list of single-node graphs
}
\description{
Explode a graph into a list of single-node graphs.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 5)
print(explode(g))
}
