% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explode.tbl_kgx}
\alias{explode.tbl_kgx}
\title{Explode a graph into a list of single-node graphs.}
\usage{
\method{explode}{tbl_kgx}(x, ...)
}
\arguments{
\item{x}{A graph object}
}
\value{
A list of single-node graphs
}
\description{
Explode a graph into a list of single-node graphs.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 5)
print(explode(g))
}
