% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nodes.neo4j_engine.R
\name{expr_to_cypher}
\alias{expr_to_cypher}
\title{This function takes an R expression and converts it to Cypher syntax.}
\usage{
expr_to_cypher(expr)
}
\arguments{
\item{expr}{An R expression}
}
\value{
A string containing the expression in Cypher syntax
}
\description{
This function takes an R expression and converts it to Cypher syntax.
}
\examples{
expr_to_cypher(~ id == "MONDO:0007525")
expr_to_cypher(~ id == "MONDO:0007525" & category == "biolink:Disease")
}
