% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_engine.R
\name{file_engine}
\alias{file_engine}
\title{Create a knowledge graph engine object from a KGX-based tsv file}
\usage{
file_engine(filename, preferences = NULL, ...)
}
\arguments{
\item{filename}{A character string indicating the filename or URL of the KGX-based tsv file.}

\item{preferences}{A named list of preferences for the engine.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{file_engine}
}
\description{
Creates a knowledge graph engine backed by a KGX-based tab-separated file. This must be a filename or URL to a \code{.tar.gz} file containing a \verb{*_nodes.tsv} and \verb{*_edges.tsv} file. If a URL is provided, the file will be downloaded to the user's current working directory.
}
\details{
Engines store preference information specifying how data are fetched and manipulated; for example,
while node \code{category} is multi-valued (nodes may have multiple categories, for example "biolink:Gene" and "biolink:NamedThing"),
typically a single category is used to represent the node in a graph, and is returned as the nodes' \code{pcategory}. A preference list of categories to use for \code{pcategory} is
stored in the engine's preferences. A default set of preferences is stored in the package for use with KGX (BioLink-compatible) graphs (see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md),
but these can be overridden by the user.

For \code{file_engine()}s, preferences are also used to set the node properties to search when using \code{search_kg()}, defaulting to regex-based searches on id, name, and description.
}
\examples{
library(tidygraph)
library(dplyr)

# Using a local MONDO KGX file (packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")
e <- file_engine(filename)
print(e$preferences)   # print the default preferences

# same search and fetch, different preferences
search_kg(e, "fibrosis", limit = 5) |>
  activate(nodes) |>
  as.data.frame() |>
  select(name, id, pcategory, category)

# prefer to set pcategory to "biolink:ThingWithTaxon" if it applies,
# followed by "biolink:NamedThing", otherwise use the first listed category.
# Additionally, only search nodes' name property.
# (Note that the MONDO KGX file does not provide multiple categories for
# nodes, so the category_priority preference has no effect here.)
e <- file_engine(filename,
                 preferences = list(category_priority = c("biolink:ThingWithTaxon", "biolink:NamedThing"),
                                    node_search_properties = c("name")))

search_kg(e, "fibrosis", limit = 5) |>
  activate(nodes) |>
  as.data.frame() |>
  select(name, id, pcategory, category)

}
\seealso{
\code{neo4j_engine()}, \code{monarch_engine()}
}
