% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_engine_check.R
\name{file_engine_check}
\alias{file_engine_check}
\title{Check availability of a file-based engine}
\usage{
file_engine_check(filename, warn = TRUE)
}
\arguments{
\item{filename}{A character string indicating the path to the file-based
engine.}

\item{warn}{A logical indicating whether to print a warning message if with
failure information if the database is not available or not
properly formatted. Default is TRUE.}
}
\value{
TRUE if the database is available and properly formatted, FALSE
otherwise.
}
\description{
Attempts to connect to the specified file-based engine. Returns FALSE if the
file is not available or not properly formatted.
}
\examples{
\dontshow{if (file_engine_check("https://kghub.io/kg-obo/sepio/2023-06-13/sepio_kgx_tsv.tar.gz")) withAutoprint(\{ # examplesIf}
print(file_engine_check("https://kghub.io/kg-obo/sepio/2023-06-13/sepio_kgx_tsv.tar.gz"))
print(file_engine_check("https://no-such-host.kghub.io/sepio_kgx_tsv.tar.gz"))
print(file_engine_check(system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")))
print(file_engine_check(system.file("extdata", "nosuch_kgx_tsv.tar.gz", package = "monarchr")))
\dontshow{\}) # examplesIf}
}
