% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_list_of_strings}
\alias{flatten_list_of_strings}
\title{Flattens and concatenates character vectors in a list}
\usage{
flatten_list_of_strings(strings)
}
\arguments{
\item{strings}{A list of character vectors.}
}
\value{
Character vector, same length as the initial list.
}
\description{
Flattens and concatenates character vectors in a list
}
\examples{
words <- list("a", NULL, c("few", "words"), NA, c("to", "try"), c(NA, NA), c("and", NA, "finally"))
flatten_list_of_strings(words)
}
