% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nodes.neo4j_engine.R
\name{generate_cypher_query}
\alias{generate_cypher_query}
\title{This function generates a Cypher query based on a set of conditions.}
\usage{
generate_cypher_query(..., limit = 10, skip = 0)
}
\arguments{
\item{...}{A set of conditions}

\item{limit}{The maximum number of results to return}
}
\value{
A string containing the Cypher query
}
\description{
This function generates a Cypher query based on a set of conditions.
}
\details{
This function takes a set of conditions and generates a Cypher query that
can be used to query the Monarch knowledge graph. The conditions should be R expressions,
and can be combined using the logical operators \code{&} (AND) and \code{|} (OR). The \code{==} operator
is used to test for equality, and the \code{\%in\%} operator is used to test for membership in a list.
Also supported are the \code{!} (NOT) operator, \code{!=}, and paretheses for grouping.
Uses tidy syntax, and multiple conditions supplied as parameters are combined with \code{AND}.

Note that expression like fetch_nodes(id \%in\% c("MONDO:0007525", "HGNC:4635")) will \emph{not} work; see query_ids().
}
\examples{
generate_cypher_query(id == "MONDO:0007525")
generate_cypher_query(id == "MONDO:0007525" & category == "biolink:Disease")
generate_cypher_query("biolink:Disease" \%in\% category | "biolink:Gene" \%in\% category)
}
