% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_association_all}
\alias{get_association_all}
\title{Retrieve All Associations for a Given Entity in a Specified Category}
\usage{
get_association_all(category, entity, limit, offset)
}
\arguments{
\item{category}{A character string specifying the association category. Example association categories are
"biolink:CorrelatedGeneToDiseaseAssociation", "biolink:DiseaseToPhenotypicQualityAssociation", etc.}

\item{entity}{A character string specifying the Monarch Initiative entity id, e.g., "MONDO:0009061", "HGNC:1884", etc.}

\item{limit}{An integer specifying the maximum number of results to return. Useful for pagination.}

\item{offset}{An integer specifying the number of results to skip. Useful for pagination.}
}
\value{
A \code{monarch_kg} object representing a graph with the associations of the specified entity.
}
\description{
This function retrieves all associations for a given entity from the Monarch Initiative API within a specified category.
The result is a \code{monarch_kg} object representing a graph with the associations.
}
\details{
This function is not exported and meant for internal use. As such it assumes that entity and category are valid identifiers, and will stop if not.
}
\examples{
# errors:
# get_association_all("sdf", "MONDO:0009061", 5, 1)
# get_association_all("biolink:CorrelatedGeneToDiseaseAssociation", "sdfsdf", 5, 1) # error

# returns monarch_kg with 5 nodes
get_association_all("biolink:CorrelatedGeneToDiseaseAssociation", "HGNC:1884", 5, 1)

}
