% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_metadata.R
\name{get_association_categories}
\alias{get_association_categories}
\title{Get association categories}
\usage{
get_association_categories(type1 = "Gene", type2 = "Disease")
}
\arguments{
\item{type1}{A character string indicating the first entity type}

\item{type2}{A character string indicating the second entity type}
}
\value{
A character vector of association categories
}
\description{
Given two entity type labels, return the association categories that exist between them
}
\details{
This function maps pairs of human-labels (e.g. "Gene" instead of "biolink:Gene") to the association categories
that connect them. Note that the order of the types does not matter, so get_association_categories("Gene", "Disease")
and get_association_categories("Disease", "Gene") will return the same result, even though the association is directional.
}
\examples{
get_association_categories("Gene", "Disease")
get_association_categories("Disease", "Gene") # same result
}
