% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_engine.R
\name{get_engine}
\alias{get_engine}
\title{Get engine}
\usage{
get_engine(g, fail_if_missing = TRUE)
}
\arguments{
\item{g}{A tbl_kgx graph.}

\item{fail_if_missing}{If TRUE, fail if there is no engine associated with the graph.}
}
\value{
A graph engine object.
}
\description{
Given a tbl_kgx graph, retrieve the last-used engine.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 5)
engine <- get_engine(g)
}
