% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_entity}
\alias{get_entity}
\title{Get entity data from the Monarch Initiative API}
\usage{
get_entity(entity = NULL)
}
\arguments{
\item{entity}{A character string indicating the entity id. Default is NULL.}
}
\value{
A tibble with the retrieved entity data, or a dummy tibble with
the correct columns but no data if the entity is NULL or not found.
}
\description{
This function retrieves data for a given entity from the Monarch Initiative API.
If no entity is provided, or if the provided entity is not found in the API,
the function returns a dummy dataframe with the appropriate columns to become nodes
in a monarch_kg.
}
\examples{
get_entity("MONDO:0005148")
get_entity() # returns an empty data tibble
get_entity("sdfsdf") # returns and empty tibble and produces a warning
}
