% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_metadata.R
\name{get_entity_types}
\alias{get_entity_types}
\title{Get entity types}
\usage{
get_entity_types()
}
\value{
A character vector of potential human-readable entity types
}
\description{
Returns a character vector of potential human-readable entity types, e.g. "Gene", "Disease", etc.
}
\details{
These are the corresponding biolink classes with the "biolink:" prefix removed (e.g. "biolink:Gene" -> "Gene")
}
\examples{
get_entity_types()
}
