% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_tilde.R
\name{\%~\%}
\alias{\%~\%}
\title{Infix regular expression match.}
\usage{
string \%~\% pattern
}
\arguments{
\item{string}{A character vector to look for matches in.}

\item{pattern}{A regular expression to look for.}
}
\value{
A logical vector indicating pattern matches.
}
\description{
An infix alias for stringr::str_detect().
}
\examples{
rownames(mtcars) \%~\% "^Merc"

}
