% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_centrality.R
\name{graph_centrality}
\alias{graph_centrality}
\title{Add centrality}
\usage{
graph_centrality(
  graph,
  fun = igraph::harmonic_centrality,
  col = "centrality",
  ...
)
}
\arguments{
\item{graph}{A graph object}

\item{fun}{The centrality function to use.
Default is \link[igraph]{harmonic_centrality}.}

\item{col}{Name of the new node attribute to store the centrality score.}

\item{...}{Additional arguments passed to the centrality function
(\code{sim_fun}).}
}
\value{
Graph object with centrality added as a new node attribute.
}
\description{
First computes of each node in a graph.
Then adds the centrality score as an node attribute.
}
\examples{
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")
g <- file_engine(filename) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")|>
          expand(categories = "biolink:Gene")
g <- graph_centrality(g)
nodes(g)$centrality
}
