% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_semsim.R
\name{graph_semsim}
\alias{graph_semsim}
\title{Add semantic similarity}
\usage{
graph_semsim(
  graph,
  fun = igraph::similarity,
  col = "similarity",
  nm = "id",
  sparse = TRUE,
  return_matrix = FALSE,
  ...
)
}
\arguments{
\item{graph}{A graph object}

\item{fun}{The similarity function to use.
Default is \link[igraph]{similarity}.}

\item{col}{Name of the new edge attribute to store the similarity score.}

\item{nm}{The node attribute to use as the matrix row/colnames.}

\item{sparse}{Return a sparse matrix instead of a dense matrix.}

\item{return_matrix}{Return the similarity matrix instead of the graph.}

\item{...}{Additional arguments passed to the similarity function
(\code{fun}).}
}
\value{
Graph object with similarity added as a new edge attribute.
}
\description{
First computes semantic similarity between all pairs of nodes in a graph.
Then adds the continuous similarity score as an edge attribute.
}
\examples{
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")
g <- file_engine(filename) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")|>
          expand(categories = "biolink:Gene")
g <- graph_semsim(g)
edges(g)$similarity
}
