% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_sparsity.R
\name{graph_sparsity}
\alias{graph_sparsity}
\title{Compute sparsity}
\usage{
graph_sparsity(x, fun = igraph::as_adjacency_matrix, ...)
}
\arguments{
\item{x}{An igraph object or a (sparse) matrix.}

\item{fun}{Function to convert graph to matrix.}

\item{...}{Arguments passed to \code{fun}.}
}
\value{
A numeric value representing the proportion of zero
values in the graph/matrix.
}
\description{
Compute sparsity (proportion of zero values) in a matrix or graph.
}
\examples{
## Using example KGX file packaged with monarchr
data(eds_marfan_kg)
g <- eds_marfan_kg |>
     fetch_nodes(query_ids = "MONDO:0007525") |>
     expand(predicates = "biolink:has_phenotype",
            categories = "biolink:PhenotypicFeature")

graph_sparsity(g)
}
