% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_edge_weights.R
\name{kg_edge_weights}
\alias{kg_edge_weights}
\title{Knowledge Graph edge weights}
\usage{
kg_edge_weights(
  graph,
  normalise = TRUE,
  encodings = list(knowledge_level = list(knowledge_assertion = 1, logical_entailment =
    1, not_provided = 0), frequency_qualifier = list(`HP:0040281` = 1, `HP:0040282` =
    0.75, `HP:0040283` = 0.5, `HP:0040284` = 0.25), negated = list(`TRUE` = -1, `FALSE` =
    0), has_total = NULL, has_quotient = numeric(), has_count = NULL, has_percentage =
    NULL, has_evidence = NULL, onset_qualifier = NULL, publications = function(x) {
    
    length(unique(x))
 }),
  fun = function(x) {
     rowSums(x, na.rm = TRUE)
 }
)
}
\arguments{
\item{graph}{A graph object}

\item{normalise}{Normalise each encoding from 0-1 by dividing by the
maximum value. Default is \code{TRUE}.}

\item{encodings}{A list of named lists of encoding values for
different edge attributes.}

\item{fun}{Function to compute edge weights with across the
numerically encoded attributes. Default is \link{rowSums}.}
}
\description{
Compute edge weights for the given \link{tbl_kgx} graph using
several pieces of categorical, ordinal, and continuous metadata.
}
\examples{
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")
g <- file_engine(filename) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")|>
          expand(categories = "biolink:Gene")
g2 <- kg_edge_weights(g)
edges(g2)$weight
}
