% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.tbl_kgx.R
\name{knit_print.tbl_kgx}
\alias{knit_print.tbl_kgx}
\title{Specialized print function for KGX graphs in knitted documents}
\usage{
\method{knit_print}{tbl_kgx}(x, ..., show = 100)
}
\arguments{
\item{x}{A \code{tbl_kgx} graph to display.}

\item{...}{Other arguments (unused).}

\item{show}{The maximum number of nodes and edges to display.}
}
\description{
Specialized print function for KGX graphs in knitted documents
}
