% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_umap.R
\name{layout_umap}
\alias{layout_umap}
\title{Layout UMAP}
\usage{
layout_umap(graph, use_3d = FALSE, prefix = "UMAP", ...)
}
\arguments{
\item{graph}{A graph object}

\item{use_3d}{Logical, whether to use a 3D layout (TRUE) or
2D layout (FALSE). Default is FALSE (2D).}

\item{prefix}{A character string prefix to add to the layout column names.}

\item{...}{Additional arguments passed to the layout function.}
}
\value{
A matrix of x and y coordinates for each node in the graph.
}
\description{
Generate a 2D or 2D layout of a graph using the UMAP algorithm.
See here for details:
\href{https://igraph.org/c/doc/igraph-Layout.html#igraph_layout_umap}{
igraph_layout_umap}
}
\examples{
set.seed(2024)
data(eds_marfan_kg)
g <- eds_marfan_kg |>
     fetch_nodes(pcategory=="biolink:Disease", limit=40) |>
     expand(predicates = "biolink:has_phenotype",
            categories = "biolink:PhenotypicFeature")|>
     tidygraph::sample_n(200) |>
     expand(categories = "biolink:Gene")
X <- layout_umap(g)
g <- graph_centrality(g)
plot(g, layout=X, node_size=centrality)
}
