% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{list_of_paths_to_basenames}
\alias{list_of_paths_to_basenames}
\title{Extracts and concatenates (possibly multiple) basenames from paths.}
\usage{
list_of_paths_to_basenames(paths)
}
\arguments{
\item{paths}{A list of character vectors representing file or url paths.}
}
\value{
The basenames as a flattened character vector.
}
\description{
Basenames are concatenated if there are multiple paths per input.
}
\examples{
paths <- list('a/b.w', c('c/d.x', 'e/f.y'), 'g/h.z')
list_of_paths_to_basenames(paths)
}
