% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_lists}
\alias{merge_lists}
\title{Recursively Merge Two Named Lists}
\usage{
merge_lists(a, b)
}
\arguments{
\item{a}{A named list to be updated.}

\item{b}{A named list containing values to override or add to \code{a}.}
}
\value{
A named list that is the result of merging \code{b} into \code{a}.
}
\description{
This function takes two named lists and merges them recursively. Values in
the second list (\code{b}) override or extend those in the first list (\code{a}).
If both \code{a} and \code{b} contain a named list at the same position, they are merged recursively.
}
\examples{
a <- list(x = list(y = 1, z = 2), foo = "bar")
b <- list(x = list(y = 42, new = 99), foo = "baz", extra = "new_value")
merge_lists(a, b)
}
