% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{monarch_api}
\alias{monarch_api}
\title{Gets response from monarch API.}
\usage{
monarch_api(url)
}
\arguments{
\item{url}{URL as a string.}
}
\value{
A monarch_api S3 class with content as R objects, the url
used for the query and the server response.
}
\description{
If a server status error is encoutered, content is set to NULL.
Otherwise it is parsed from json to R objects.
}
\examples{
url <- "https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene\%3A8314?rows=100"
resp <- monarch_api(url)
}
