% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_engine.R
\name{monarch_engine}
\alias{monarch_engine}
\title{Create a knowledge graph engine object backed by the public Monarch Neo4j
instance}
\usage{
monarch_engine(
  url = c("https://neo4j.monarchinitiative.org", "http://neo4j.monarchinitiative.org",
    "https://neo4j.monarchinitiative.org:7473",
    "http://neo4j.monarchinitiative.org:7473"),
  api_url = "https://api.monarchinitiative.org/v3/api",
  preferences = NULL,
  timeout = 2,
  ...
)
}
\arguments{
\item{url}{(Optional) May be specified to override the default Monarch Neo4j
URL. If given a vector, each will be tried in sequence; if a URL
times out (see timeout) or fails, the next is tried.}

\item{api_url}{(Optional) May be specified to override the default Monarch
API URL (specifying the location of the \verb{/search} endpoint
used by \code{search_nodes()}).}

\item{preferences}{A named list of preferences for the engine.}

\item{timeout}{Number of seconds to wait before trying the next url.}

\item{...}{Additional arguments passed to \code{neo2R::startGraph()}.}
}
\value{
An object of class \code{monarch_engine}
}
\description{
Creates a knowledge graph engine backed by the publicly hosted Monarch Neo4j
database, used to fetch nodes and edges from the database as local
graph objects.
}
\details{
Engines store preference information specifying how data are fetched and
manipulated; for example,
while node \code{category} is multi-valued (nodes may have multiple categories,
for example "biolink:Gene" and "biolink:NamedThing"),
typically a single category is used to represent the node in a graph, and is
returned as the nodes' \code{pcategory}. A preference list of categories to use
for \code{pcategory} is
stored in the engine's preferences. A default set of preferences is stored in the package for use with this and other KGX (BioLink-compatible) graphs (see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md),
but these can be overridden by the user.

The \code{monarch_engine()} overrides \code{search_nodes()} to use the Monarch search API, so setting \code{node_search_properties} in the preferences will not affect the search behavior. To use regex-based searching with the Monarch Neo4j instance, use \code{neo4j_engine()} instead and specify the Monarch Neo4j URL (https://neo4j.monarchinitiative.org).
}
\examples{
\dontshow{if (monarch_engine_check()) withAutoprint(\{ # examplesIf}
library(tidygraph)
library(dplyr)

monarch <- monarch_engine()
res <- monarch |> fetch_nodes(query_ids = c("MONDO:0007522", "MONDO:0007947"))
print(res)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{file_engine()}, \code{neo4j_engine()}
}
