% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_engine_check.R
\name{monarch_engine_check}
\alias{monarch_engine_check}
\title{Check availability of Monarch Initiative database}
\usage{
monarch_engine_check(url, username = NA, password = NA)
}
\value{
TRUE if the database is available and properly formatted, FALSE otherwise.
}
\description{
Attempts to connect to the Monarch Iniative Neo4J database and run a query to see if it is properly formatted. Returns FALSE if the database is not available or not properly formatted.
}
\examples{
print(monarch_engine_check())

}
