% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query.neo4j_engine.R
\name{neo2r_to_kgx}
\alias{neo2r_to_kgx}
\title{Process neo2R cypher to tbl_kgx}
\usage{
neo2r_to_kgx(res, engine)
}
\arguments{
\item{res}{The result from neo2R::cypher with result = "graph"}

\item{engine}{The engine to attach to the returned graph}
}
\value{
A tbl_kgx
}
\description{
Given a result from neo2R::cypher returning KGX-formatted nodes and edges,
parse the result to generate a tbl_kgx object, attaching the provided engine.
}
