% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_engine_check.R
\name{neo4j_engine_check}
\alias{neo4j_engine_check}
\title{Check if a neo4j database is available and properly formatted}
\usage{
neo4j_engine_check(url, username = NA, password = NA)
}
\arguments{
\item{url}{A character string indicating the URL of the neo4j database.}

\item{username}{A character string indicating the username for the neo4j database (if needed).}

\item{password}{A character string indicating the password for the neo4j database (if needed).}
}
\value{
TRUE if the database is available and properly formatted, FALSE otherwise.
}
\description{
Attempts to connect to the specified Neo4J database and run a query to see if it is properly formatted. Returns FALSE if the database is not available or not properly formatted.
}
\examples{
print(neo4j_engine_check("https://neo4j.monarchinitiative.org"))
print(neo4j_engine_check("https://no-such-db.monarchinitiative.org"))

}
