% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes}
\alias{nodes}
\title{Get graph nodes table.}
\usage{
nodes(x, ...)
}
\arguments{
\item{x}{A graph object}
}
\value{
A tibble with the nodes of the graph
}
\description{
Get graph nodes table.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 1)
print(nodes(g))
print(edges(g))
}
