% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes.tbl_kgx}
\alias{nodes.tbl_kgx}
\title{Get tbl_kgx graph nodes table.}
\usage{
\method{nodes}{tbl_kgx}(x, ...)
}
\description{
Get tbl_kgx graph nodes table.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 1)
print(nodes(g))
}
