% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{null_to_na}
\alias{null_to_na}
\title{Replace NULL with NA}
\usage{
null_to_na(x, na_type_fun = as.character)
}
\arguments{
\item{x}{A variable of any type. This is the variable that will be checked
for NULL and possibly replaced with NA.}

\item{na_type_fun}{A function that converts NA to a specific type.
This function should take NA as an input and return NA of a certain type.
By default, this is \code{as.character}, which will return NA_character_.
Other possibilities are \code{as.integer}, \code{as.numeric}, etc.}
}
\value{
The original variable if it is not NULL, or NA of a specific type
if the original variable is NULL.
}
\description{
This function checks if an input is NULL, and if so, replaces it with NA
of a specific type. If the input is not NULL, the function simply returns the input.
}
\examples{
null_to_na(NULL) # returns a character NA
null_to_na(NULL, as.integer) # returns an integer NA
null_to_na("Not null") # returns "Not null"
}
