% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cols.R
\name{order_cols}
\alias{order_cols}
\title{Set edge/row data column order according to most recent engine preferences}
\usage{
order_cols(g)
}
\arguments{
\item{g}{A \code{tbl_kgx} graph.}
}
\value{
The input graph with ordered node/edge columns.
}
\description{
Set edge/row data column order according to most recent engine preferences
}
