% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{print.monarch_api}
\alias{print.monarch_api}
\title{Print method for monarch_api.}
\usage{
\method{print}{monarch_api}(x, depth = 2, ...)
}
\arguments{
\item{x}{A monarch_api S3 Object.}

\item{depth}{An integer indicating how many levels of content to show.}

\item{...}{Further options for print.}
}
\value{
A summary of x and, invisibly, x.
}
\description{
Print method for monarch_api.
}
