% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.file_engine.R
\name{print.summary.file_engine}
\alias{print.summary.file_engine}
\title{Print a file-engine summary}
\usage{
\method{print}{summary.file_engine}(x, ...)
}
\arguments{
\item{x}{A \code{summary.file_engine} object.}

\item{...}{Other parameters (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a file-engine summary
}
