% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.neo4j_engine.R
\name{print.summary.neo4j_engine}
\alias{print.summary.neo4j_engine}
\title{Print a neo4j-engine summary}
\usage{
\method{print}{summary.neo4j_engine}(x, ...)
}
\arguments{
\item{x}{A \code{summary.neo4j_engine} object.}

\item{...}{Other parameters (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a neo4j-engine summary
}
