% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tbl_kgx.R
\name{print.tbl_kgx}
\alias{print.tbl_kgx}
\title{Print function for KGs.}
\usage{
\method{print}{tbl_kgx}(x, suppress_prefix = NULL, ...)
}
\arguments{
\item{suppress_prefix}{A character vector of prefixes to suppress in the printed output}

\item{...}{Additional arguments to be passed to the print method for tbl_graph}

\item{g}{A graph}
}
\value{
The graph
}
\description{
Provides basic features over tidygraph print.
}
