% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_graph.R
\name{query_graph}
\alias{query_graph}
\title{Query Graph}
\usage{
query_graph(query, parameters = NULL, ...)
}
\arguments{
\item{query}{A string representing the cypher query.}

\item{parameters}{Parameters for the cypher query. Default is NULL.}

\item{...}{Additional arguments passed to the function (unused).}
}
\value{
A tbl_kgx graph.
}
\description{
This function takes a cypher query and returns a tbl_kgx graph. It retrieves the graph connection and knowledge graph preferences from the parent environment.
It then executes the cypher query and stitches the vectors. It creates a data frame for nodes and edges and adds all other node and edge properties as columns.
It also computes a priority category based on a preference list. Finally, it creates a tbl_kgx graph from the nodes and edges data frames.
}
\examples{
query <- "MATCH (s) -[p]- (o) return s, p, o LIMIT 2"
parameters <- NULL
g <- query_graph(query, parameters)

query = "MATCH (n) WHERE n.id IN $ids RETURN n"
parameters = list(ids = ids)
g <- query_graph(query, parameters)
}
