% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ids.neo4j_engine.R
\name{query_ids.neo4j_engine}
\alias{query_ids.neo4j_engine}
\title{Query IDs}
\usage{
\method{query_ids}{neo4j_engine}(engine, ids)
}
\arguments{
\item{engine}{A graph engine object.}

\item{ids}{A vector of IDs to be queried.}
}
\value{
A tbl_kgx graph.
}
\description{
Given a vector of id values, returns a tbl_kgx graph with corresponding nodes from the knowledge graph. No edges are returned.
}
\examples{
ids <- c("MONDO:0015780", "HP:0001903", "HGNC:2890")
g <- query_ids(ids)
}
