% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{roll}
\alias{roll}
\title{Internal function for rolling up and down}
\usage{
roll(
  column,
  fun = c,
  include_self = TRUE,
  predicates = "biolink:subclass_of",
  direction = "up",
  ...
)
}
\arguments{
\item{column}{The node column to draw rollup or rolldown information from.}

\item{fun}{The aggregation function to use when rolling up or down. Default is \code{c}}

\item{include_self}{Whether to include each nodes' value in \code{column} in the rollup/rolldown for that node.}

\item{predicates}{A vector of relationship predicates (nodes in g are subjects in the KG), indicating which edges to consider in the rollup/rolldown. Should be transitive; default \code{biolink:subclass_of}}

\item{direction}{Whether to roll up or down.}

\item{...}{Other parameters (unused)}
}
\description{
Internal function for rolling up and down
}
