% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_kg.monarch_engine.R
\name{search_kg.monarch_engine}
\alias{search_kg.monarch_engine}
\title{Search the Monarch Initiative}
\usage{
\method{search_kg}{monarch_engine}(engine, query, category = NULL, limit = 10, ...)
}
\arguments{
\item{engine}{A Monarch engine object.}

\item{query}{A character string representing the query term. Example terms:
"Cystic Fibrosis", "CYP6B", "swelling of joints".}

\item{category}{A character string indicating a single entity category in which to search for the query term.
Defaults to NULL, to search in any category.}

\item{limit}{An integer indicating the maximum number of search results to return. Defaults to 10.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{tbl_kgx} graph object containing the search results as nodes, with no edges.
}
\description{
This function uses the Monarch Initiative API to search for entities matching
a given query string within a specified category, returing a graph containg just the nodes in the search results.
}
\details{
The returned graph will contain only nodes with no associations between them, even if they exist in the Monarch database.
}
\examples{
e <- monarch_engine()
search(e, "Cystic Fibrosis", category = "biolink:Disease", limit = 5)

}
