% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{seed_graph}
\alias{seed_graph}
\title{Create a Single-Node Graph from a Monarch Initiative Entity ID}
\usage{
seed_graph(entity_id = NULL)
}
\arguments{
\item{entity_id}{A character string indicating the Monarch Initiative entity id,
e.g. "MONDO:0009061" or "HGNC:1884". If NULL (default), an empty graph is returned.}
}
\value{
A \code{monarch_kg} graph object representing the entity, or an empty graph
if the entity is not found.
}
\description{
This function retrieves data for a given entity from the Monarch Initiative API and
uses it to create a graph with a single node and no edges. If the entity is not found,
it returns an empty graph.
}
\examples{
seed_graph("MONDO:0009061") # returns a monarch_kg object with one node corresponding to the entity
seed_graph() # returns an empty graph
seed_graph("sdflkjsdf")  # returns an empty graph and gives a warning
}
