% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_nodes.R
\name{select_nodes}
\alias{select_nodes}
\title{Select nodes from a graph}
\usage{
select_nodes(graph, ...)
}
\arguments{
\item{graph}{A graph object from which nodes should be selected.
The specific class of this object determines the method that will be used.}

\item{...}{Further arguments defining the conditions for selecting nodes.
These are passed to the specific \code{select_nodes} method for the class of \code{graph}.}
}
\value{
A graph with a subset of nodes selected.
}
\description{
This is the generic function for \code{select_nodes}, which selects nodes from a graph
based on specific conditions. The conditions are specified through the \code{...} argument.
The actual implementation of the function depends on the class of the \code{graph} object,
and should be defined in specific methods for the \code{select_nodes} function.
}
