% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R, R/set_engine.tbl_kgx.R
\name{set_engine}
\alias{set_engine}
\title{Set the engine for a graph. See warning in details.}
\usage{
set_engine(g, engine)

set_engine(g, engine)
}
\arguments{
\item{g}{A tbl_kgx graph.}

\item{engine}{An engine object.}
}
\value{
A tbl_kgx graph.

A tbl_kgx graph.
}
\description{
Sets a given graph's engine to a given engine object and returns the graph.
WARNING: changing the backing engine for a graph dynamically is not
yet fully supported or tested.

Sets a given graph's engine to a given engine object and returns the graph.
WARNING: changing the backing engine for a graph dynamically is not
yet fully supported or tested.
}
\examples{
# Using example KGX file packaged with monarchr
data(eds_marfan_kg)

g <- eds_marfan_kg |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

other_engine <- eds_marfan_kg # this could be a different file-engine (see `file_engine()`)
g <- set_engine(g, other_engine)

print(get_engine(g))

# Using example KGX file packaged with monarchr
data(eds_marfan_kg)

g <- eds_marfan_kg |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

other_engine <- eds_marfan_kg # this could be a different file engine (see `file_engine()`)
g <- set_engine(g, other_engine)

print(get_engine(g))

}
