% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity.R
\name{sparsity}
\alias{sparsity}
\title{Compute sparsity}
\usage{
sparsity(x, fun = igraph::as_adjacency_matrix, ...)
}
\arguments{
\item{x}{An igraph object or a (sparse) matrix.}

\item{fun}{Function to convert graph to matrix.}

\item{...}{Arguments passed to \code{fun}.}
}
\value{
A numeric value representing the proportion of zero
values in the graph/matrix.
}
\description{
Compute sparsity (proportion of zero values) in a matrix or graph.
}
\examples{
g <- make_ring(10)
sparsity(g)
}
