% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query.neo4j_engine.R
\name{stitch_vectors}
\alias{stitch_vectors}
\title{Stitch Vectors}
\usage{
stitch_vectors(x)
}
\arguments{
\item{x}{An element to be stitched.}
}
\value{
A list of single-character vectors.
}
\description{
Neo4j array results come back as lists of length-1 character vectors;
these need to be stitched together to length-N character vectors.
We leave these in a list element to distinguish sets of size 1 from
scalar strings
}
