% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood_edges.neo4j_engine.R
\name{summarize_neighborhood_edges.neo4j_engine}
\alias{summarize_neighborhood_edges.neo4j_engine}
\title{Summarize neighborhood edges}
\usage{
\method{summarize_neighborhood_edges}{neo4j_engine}(e, g, direction = "both")
}
\arguments{
\item{e}{A graph engine object}

\item{g}{A graph}

\item{direction}{The direction of edges to include in the neighborhood}
}
\value{
A tbl_kgx graph
}
\description{
Summarizes the relationships in the neighborhood of a given graph. Specifically,
this function retrieves counts of different predicates connecting nodes in the given graph to
nodes in the larger KG. Counts are broken down by source and target node categories.
Note that the number of relationships returned may be larger than the number of nodes they connect to.
}
\examples{
e <- neo4j_engine()
g <- monarch_search("fanconi anemia")
summarize_neighborhood_edges(e, g, direction = "both")

}
