% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood_edges.R
\name{summarize_neighborhood_edges}
\alias{summarize_neighborhood_edges}
\title{Summarize neighborhood edges}
\usage{
summarize_neighborhood_edges(obj, ...)
}
\arguments{
\item{e}{A graph engine object (optional - the recent engine used by the graph will be used if not supplied)}

\item{g}{A graph}

\item{direction}{The direction of edges to include in the neighborhood}
}
\value{
A tbl_kgx graph
}
\description{
Summarizes the relationships in the neighborhood of a given graph. Specifically,
this function retrieves counts of different predicates connecting nodes in the given graph to
nodes in the larger KG. Counts are broken down by source and target node categories.
Note that the number of relationships returned may be larger than the number of nodes they connect to.
}
\examples{
\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
g <- monarch_engine() |>
  search_nodes("fanconi anemia", limit = 5) |>
  summarize_neighborhood_edges(direction = "both")
\dontshow{\}) # examplesIf}
}
