% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood.R
\name{summarize_neighborhood_nodes}
\alias{summarize_neighborhood_nodes}
\title{Summarize neighborhood nodes}
\usage{
summarize_neighborhood_nodes(g, direction = "both")
}
\arguments{
\item{g}{A graph}

\item{direction}{The direction of edges to include in the neighborhood}
}
\value{
A tbl_kgx graph
}
\description{
Summarizes the nodes in the neighborhood of a given graph. Specifically,
this function retrieves counts of different node categories in the neighborhood of the given graph.
}
\examples{
g <- monarch_search("fanconi anemia")
summarize_neighborhood_nodes(g)

}
