% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.file_engine.R
\name{summary.file_engine}
\alias{summary.file_engine}
\title{Summarize contents of a KGX-file-based KG engine}
\usage{
\method{summary}{file_engine}(object, ..., quiet = FALSE)
}
\arguments{
\item{object}{A \code{file_engine} object}

\item{...}{Other parameters (not used)}

\item{quiet}{Logical, whether to suppress printing of the summary}
}
\value{
A classed list of data frames and named lists.
}
\description{
Given a KGX file-based KG engine, provides summary information in the form of
node counts, category counts across nodes, relationship type counts, and available properties.
The returned summary object prints a readable console report and also contains
data frames with this information. Also returned are \code{cats}, \code{preds}, and
\code{props} entries, containing lists of available
categories/predicates/properties for convenient auto-completion in RStudio.
}
\details{
When applied to a \code{file_engine}, also included are node-specific and edge-specific properties.
}
\examples{
# Using example KGX file packaged with monarchr
data(eds_marfan_kg)

# prints a readable summary and returns a list of dataframes
res <- eds_marfan_kg |> summary()
print(res)
}
