% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.neo4j_engine.R
\name{summary.neo4j_engine}
\alias{summary.neo4j_engine}
\title{Summarize contents of a Neo4j KG engine}
\usage{
\method{summary}{neo4j_engine}(object, ..., quiet = FALSE)
}
\arguments{
\item{object}{A \code{neo4j_engine} object}

\item{...}{Other parameters (not used)}

\item{quiet}{Logical, whether to suppress printing of the summary}
}
\value{
A classed list of data frames and named lists.
}
\description{
Given a Neo4j based KG engine, provides summary information in the form of
node counts, category counts across nodes, relationship type counts, and
available properties.
The returned summary object prints a readable console report and also
contains
data frames with this information. Also returned are \code{cats}, \code{preds}, and
\code{props} entries, containing lists of available
categories/predicates/properties for convenient auto-completion in RStudio.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# prints a readable summary and returns a list of dataframes
stats <- monarch_engine() |> summary()
print(stats)
\dontshow{\}) # examplesIf}
}
