% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitive_closure.R
\name{transitive_closure}
\alias{transitive_closure}
\title{Compute transitive closure over a predicate.}
\usage{
transitive_closure(g, predicate = "biolink:subclass_of")
}
\value{
Graph with transitive edges added.
}
\description{
Computes the transitive closure of a graph, treating the specified
predicate as transitive. Resulting edge predicates will the be the
same, but have primary_knowledge_source set to transitive_\if{html}{\out{<predicate>}}.
}
\examples{
engine <- file_engine(system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr"))

engine |> fetch_nodes(name == "Tall stature") |>
  expand_n(predicates = "biolink:subclass_of", direction = "out", n = 3) |>  # get 2 levels of ancestors
  activate(edges) |>
  filter(primary_knowledge_source == "infores:upheno") |>
  transitive_closure(predicate = "biolink:subclass_of") |>
  plot(edge_color = primary_knowledge_source)

}
\seealso{
\code{\link[=roll_up]{roll_up()}}, \code{\link[=transfer]{transfer()}}, \code{\link[=descendants]{descendants()}}, \code{\link[=ancestors]{ancestors()}}
}
