% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_genes_assoc_w_disease}
\alias{bioentity_genes_assoc_w_disease}
\title{Get genes associated with a disease}
\usage{
bioentity_genes_assoc_w_disease(disease)
}
\arguments{
\item{disease}{An id for a disease, like a MONDO ID.}
}
\description{
mimics https://api.monarchinitiative.org/api/bioentity/disease/MONDO%3A0006486/genes/?fetch_objects=true&rows=100
}
