% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_genes_assoc_w_disease}
\alias{bioentity_genes_assoc_w_disease}
\title{Get genes associated with a disease}
\usage{
bioentity_genes_assoc_w_disease(disease, rows = 100, start = 0,
  fetch_objects = FALSE)
}
\arguments{
\item{disease}{An id for a disease, like a MONDO ID.}

\item{rows}{Number of rows of results to fetch.}

\item{start}{The starting row of the results to fetch.}

\item{fetch_objects}{Whether or not to fetch objects at the top level of the json response.}
}
\value{
A list of (a list of information, monarch_api S3 class).
}
\description{
mimics https://api.monarchinitiative.org/api/bioentity/disease/MONDO%3A0006486/genes/?fetch_objects=true&rows=100
}
\examples{
disease <- "MONDO:0006486" # uveal melanoma
bioentity_genes_assoc_w_disease(disease)
}
