% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{bioentity_homologs}
\alias{bioentity_homologs}
\title{Get homologs for a gene.}
\usage{
bioentity_homologs(gene)
}
\arguments{
\item{gene}{A valid monarch initiative gene id.}
}
\value{
A list of (tibble of homolog information, monarch_api S3 class).
}
\description{
Replicates info in view: https://api.monarchinitiative.org/api/#!/bioentity/get_gene_homolog_associations
}
\details{
The monarch_api class is included in return mainly for debugging.

https://api.monarchinitiative.org/api/bioentity/gene/NCBIGene%3A8314/homologs/?rows=100&fetch_objects=true
}
\examples{
gene <-"NCBIGene:8314"
bioentity_homologs(gene)
}
