% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{build_monarch_url}
\alias{build_monarch_url}
\title{Builds URL for a monarch GET request.}
\usage{
build_monarch_url(path, query = NULL)
}
\arguments{
\item{path}{A path to the monarch resource to use.}

\item{query}{A list of url parameter settings. TRUE FALSE set to
"true" "false".}
}
\value{
URL as a string.
}
\description{
Builds URL for a monarch GET request.
}
\examples{
url <- build_monarch_url(path = list("/api/bioentity/gene", "NCBIGene\%3A8314"),
                         query = list(rows = 100, fetch_objects = "true",
                          format = "json"))
url <- build_monarch_url(path = list("/api/bioentity/gene", "NCBIGene\%3A8314"),
                         query = list(rows = 100, fetch_objects = TRUE,
                         unselect_evidence=FALSE, format = "json"))
url <- build_monarch_url(path = list("/api/bioentity/gene", "NCBIGene\%3A8314"),
                         query = list(rows = 100, fetch_objects = TRUE,
                         unselect_evidence=NULL, format = "json"))
}
