% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_monarch_url}
\alias{build_monarch_url}
\title{Builds URL for a monarch GET request.}
\usage{
build_monarch_url(path, query = NULL)
}
\arguments{
\item{path}{A path as a list to the monarch resource to use.}

\item{query}{A list of url parameter settings. TRUE FALSE set to
"true" "false".

The path needs to be safely encoded.}
}
\value{
URL as a safely encoded string.
}
\description{
Builds URL for a monarch GET request.
}
\examples{
m_path <- "/api/bioentity/gene"
gene <- "NCBIGene\%3A8314"
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = "true",
                         format = "json"))
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = TRUE,
                         unselect_evidence=FALSE,
                         format = "json"))
url <- build_monarch_url(path = list(m_path, gene),
                         query = list(rows = 100,
                         fetch_objects = TRUE,
                         unselect_evidence=NULL,
                         format = "json"))
}
\seealso{
\link[utils]{URLencode}
}
