% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query.R
\name{cypher_query}
\alias{cypher_query}
\title{Execute a Cypher Query}
\usage{
cypher_query(engine, ...)
}
\arguments{
\item{...}{Additional arguments passed to the function.}

\item{query}{A string representing the Cypher query.}

\item{parameters}{A list of parameters for the Cypher query. Default is an empty list.}
}
\value{
The result of the Cypher query as a tbl_kgx graph.
}
\description{
This function takes a Cypher query and parameters, executes the query using the given engine, and returns the result as a tbl_kgx graph.
}
\examples{
query <- "MATCH (n) RETURN n LIMIT 10"
ids <- c("MONDO:0007525", "MONDO:0020066", "MONDO:0034021")
parameters <- list(ids = ids)
result <- cypher_query(query, parameters)
}
