% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query.R
\name{cypher_query_df}
\alias{cypher_query_df}
\title{Execute a Cypher Query and Return a Data Frame}
\usage{
cypher_query_df(query, parameters = list(), ...)
}
\arguments{
\item{query}{A string representing the Cypher query.}

\item{parameters}{A list of parameters for the Cypher query. Default is an empty list.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A data frame containing the result of the Cypher query.
}
\description{
This function takes a Cypher query and parameters, executes the query on the Monarch knowledge graph, and returns the result as a data frame.
}
\examples{
query <- "MATCH (n) RETURN n LIMIT 10"
ids <- c("MONDO:0007525", "MONDO:0020066", "MONDO:0034021")
parameters <- list(ids = ids)
result <- cypher_query_df(query, parameters)
}
