% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edges}
\alias{edges}
\title{Get graph edges table.}
\usage{
edges(graph, ...)
}
\arguments{
\item{graph}{Input graph}

\item{...}{Other options (unused)}
}
\value{
A tibble with the edges of the graph
}
\description{
Get graph edges table.
}
\examples{
# (using the example KGX file packaged with monarchr)
data(eds_marfan_kg)

g <- eds_marfan_kg |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

print(edges(g))
}
