% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edges}
\alias{edges}
\title{Get graph edges table.}
\usage{
edges(g, ...)
}
\arguments{
\item{g}{A tbl_kgx graph.}
}
\arguments{
\item{x}{Input graph}

\item{...}{Other options (unused)}
}
\value{
A tibble with the edges of the graph
}
\description{
Get graph edges table.
}
\examples{
# (using the MONDO KGX file packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")

g <- file_engine(filename) |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

print(edges(g))
}
