% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edges}
\alias{edges}
\title{Get graph edges table.}
\usage{
edges(x, ...)
}
\arguments{
\item{x}{Input graph}

\item{...}{Other options (unused)}
}
\value{
A tibble with the edges of the graph
}
\description{
Get graph edges table.
}
\examples{
g <- monarch_search("fanconi anemia", limit = 1)
print(nodes(g))
print(edges(g))
}
