% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graph.R
\name{example_graph}
\alias{example_graph}
\title{Return an example set of nodes from a KG engine.}
\usage{
example_graph(engine, ...)
}
\arguments{
\item{engine}{A KG engine object}

\item{...}{Other parameters (not used)}
}
\value{
A tbl_kgx graph
}
\description{
Given a KG engine, returns a graph representing the diversity
of node categories and edge predicates for browsing. The returned graph is guaranteed to
contain at least one node of every category, and at least one edge of every
predicate. No other guarantees are made: the example graph is not minimal
to satisfy these criteria, it is not random or even pseudo-random, and it
may not be connected.
}
\examples{
# Using example KGX file packaged with monarchr
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")

# prints a readable summary and returns a list of dataframes
g <- file_engine(filename) |> example_graph()
print(g)

\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# prints a readable summary and returns a list of dataframes
g <- monarch_engine() |> example_graph()
print(g)
\dontshow{\}) # examplesIf}
}
