% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand a monarch_kg graph.}
\usage{
expand(graph, types = NULL, verbose = FALSE)
}
\arguments{
\item{graph}{A monarch_kg graph object.}

\item{types}{A character vector of entity types to expand to. If NULL (default), all types
will be expanded to.}

\item{verbose}{A logical indicating whether to print progress messages. Default is FALSE.}
}
\value{
A monarch_kg graph object.
}
\description{
Expand a monarch_kg graph by adding nodes linked to the selected node set of the given types.
All selected nodes will be expanded from, even if they are of different types. For example,
if the selected nodes include a disease and a gene, and the types include "Gene", then all
genes linked to the selected disease and all genes linked to the selected gene will be added
to the graph.
}
\examples{
# Expand a monarch_kg graph to all genes linked to the selected disease
graph <- seed_graph("MONDO:0005148")
graph <- expand(graph, types = "Gene")
}
