% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Fetch additional knowledge graph edges connected to a query graph}
\usage{
expand(
  engine = NULL,
  graph,
  direction = "both",
  predicates = NULL,
  result_categories = NULL,
  transitive = FALSE,
  drop_unused_query_nodes = FALSE
)
}
\arguments{
\item{engine}{(Optional) An engine to use for fetching query graph edges.}

\item{graph}{A query \code{tbl_kgx()} graph to query from.}

\item{direction}{The direction of associations to fetch. Can be "in", "out", or "both". Default is "both".}

\item{predicates}{A vector of relationship predicates (nodes in g are subjects in the KG), indicating which edges to consider in the neighborhood. If NULL (default), all edges are considered.}

\item{result_categories}{A vector of node categories, indicating which nodes in the larger KG may be fetched. If NULL (default), all nodes in the larger KG are will be fetched.}

\item{transitive}{If TRUE, include transitive closure of the neighborhood. Default is FALSE. Useful in combination with predicates like \code{biolink:subclass_of}.}

\item{drop_unused_query_nodes}{If TRUE, remove query nodes from the result, unless they are at the neighborhood boundary, i.e., required for connecting to the result nodes. Default is FALSE.}
}
\value{
A \code{tbl_kgx()} graph
}
\description{
Given an optional KG engine (e.g. a \code{file_engine()},
\code{neo4j_engine()}, or \code{monarch_engine()}) and a query \code{tbl_kgx()} graph, fetches additional nodes and edges
from the KG, expanding the query graph according to specific criteria. If the first parameter is an engine, that
engine is used; if the first parameter is a query graph, the most recent engine associated with the graph is used.
}
\examples{
\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Using Monarch (hosted)
phenos <- monarch_engine() |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype", result_categories = "biolink:PhenotypicFeature")

print(phenos)


\dontshow{\}) # examplesIf}
## Using local MONDO KGX file (packaged with monarchr)
phenos <- file_engine(system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype", result_categories = "biolink:PhenotypicFeature")

print(phenos)



\dontshow{if (file_engine_check("https://kghub.io/kg-obo/mondo/2024-03-04/mondo_kgx_tsv.tar.gz")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Using MONDO KGX file (remote) as an example
phenos <- file_engine("https://kghub.io/kg-obo/mondo/2024-03-04/mondo_kgx_tsv.tar.gz") |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype", result_categories = "biolink:PhenotypicFeature")

print(phenos)

file.remove("mondo_kgx_tsv.tar.gz") # cleanup - remove the downloaded file
\dontshow{\}) # examplesIf}
}
