% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explode}
\alias{explode}
\title{Explode a graph into a list of single-node graphs}
\usage{
explode(graph, ...)
}
\arguments{
\item{graph}{A tbl_kgx graph.}

\item{...}{Other options (unused)}
}
\value{
A list of tbl_kgx graphs.
}
\description{
Explode a graph into a list of single-node graphs
}
\examples{
# (using the MONDO KGX file packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")

g <- file_engine(filename) |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

print(explode(g))
}
