% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarchr.R
\name{extract_matching_phrases_from_lists}
\alias{extract_matching_phrases_from_lists}
\title{Finds phrases from list of things, like a list of dataframes.}
\usage{
extract_matching_phrases_from_lists(things, phrase)
}
\arguments{
\item{things}{Lists. Maybe lists of dataframes.}

\item{phrase}{An anchor phrase. (maybe to fetch cell matching that from df).}
}
\value{
List matching the phrase
}
\description{
Wherever the phrase exists, it is extracted. Convenient for list of
dataframes in which we know the phrase occurs somewhere but we don't
no the column, or don't care about preserving the source column name.
}
\examples{
#evidence <- extract_matching_phrases_from_lists(homs$evidence_graph.nodes, 'evidence')
}
