% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nodes.R
\name{fetch_nodes}
\alias{fetch_nodes}
\title{Fetch nodes from a graph}
\usage{
fetch_nodes(engine, ..., query_ids = NULL)
}
\arguments{
\item{engine}{A graph engine object}

\item{...}{A set of conditions identifying the nodes to fetch, only used if query_ids is NULL}

\item{query_ids}{A character vector of identifiers to fetch}
}
\value{
A tbl_kgx object containing the nodes
}
\description{
This function fetches nodes from a graph based on a set of conditions.
}
\examples{
e <- neo4j_engine()
g <- fetch_nodes(e, id == "MONDO:0007525")
g <- fetch_nodes(e, "biolink:Disease" \%in\% category | "biolink:Gene" \%in\% category)
}
