% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_engine.R
\name{file_engine}
\alias{file_engine}
\title{File-based knowledge graph engine}
\usage{
file_engine(filename, preferences = NULL, ...)
}
\arguments{
\item{filename}{The filename or URL to use.}

\item{preferences}{Override default preferences.}

\item{...}{Other parameter (unused).}
}
\description{
Returns a Knowledge Graph engine backed by a file hosted by kghub.io. Must be a \verb{_tsv.tar.gz} file containing nodes and edges tab-separated files in KGX format. See https://kghub.io for details.
}
