% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_engine.R
\name{file_engine}
\alias{file_engine}
\title{Create a knowledge graph engine object from a KGX-based tsv file}
\usage{
file_engine(filename, preferences = NULL, ...)
}
\arguments{
\item{filename}{A character string indicating the filename or URL of the KGX-based tsv file.}

\item{preferences}{A named list of preferences for the engine.}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class \code{file_engine}
}
\description{
Creates a knowledge graph engine backed by a KGX-based tab-separated file. This must be a filename or URL to a \code{.tar.gz} file containing a \verb{*_nodes.tsv} and \verb{*_edges.tsv} file. If a URL is provided, the file will be downloaded to the user's current working directory.
}
\details{
Engines store preference information specifying how data are fetched and manipulated; for example,
while node \code{category} is multi-valued (nodes may have multiple categories, for example "biolink:Gene" and "biolink:NamedThing"),
typically a single category is used to represent the node in a graph, and is returned as the nodes' \code{pcategory}. A preference list of categories to use for \code{pcategory} is
stored in the engine's preferences. A default set of preferences is stored in the package for use with KGX (BioLink-compatible) graphs (see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md),
but these can be overridden by the user.
}
\examples{
library(tidygraph)
library(dplyr)

# Using a local MONDO KGX file (packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")
engine <- file_engine(filename)

res <- engine |> fetch_nodes(query_ids = c("MONDO:0007522", "MONDO:0007947"))
print(res)

}
\seealso{
\code{neo4j_engine()}, \code{monarch_engine()}
}
