% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.tbl_kgx.R
\name{knit_print.tbl_kgx}
\alias{knit_print.tbl_kgx}
\title{Specialized print function for KGX graphs in knitted documents}
\usage{
\method{knit_print}{tbl_kgx}(graph, show = 100, ...)
}
\arguments{
\item{graph}{A \code{tbl_kgx} graph to display.}

\item{show}{The maximum number of rows and edges to display.}

\item{...}{Other arguments (unused).}
}
\description{
Specialized print function for KGX graphs in knitted documents
}
