% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kgx.R
\name{load_kgx}
\alias{load_kgx}
\title{Load a graph from a KGX-formatted .tar.gz file.}
\usage{
load_kgx(filename, attach_engine = NULL, ...)
}
\arguments{
\item{filename}{File to the graph from. Must end in .tar.gz and conform to KGX specification (see description).}

\item{attach_engine}{An engine to attach to the graph (optional).}

\item{...}{Other parameters (unused)}
}
\value{
A \code{tbl_kgx} graph.
}
\description{
Given a KGX-formatted tabular KG
(see https://github.com/biolink/kgx/blob/master/specification/kgx-format.md)
loads it as a graph.
}
\examples{
\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
phenos <- monarch_engine() |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")

save_kgx(phenos, "phenos.tar.gz")

# when loading the graph, we can optionally attach an engine
loaded_phenos <- load_kgx("phenos.tar.gz", attach_engine = monarch_engine())

loaded_phenos

# cleanup saved file
file.remove("phenos.tar.gz")
\dontshow{\}) # examplesIf}
}
