% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_engine.R
\name{monarch_engine}
\alias{monarch_engine}
\title{monarch_engine}
\usage{
monarch_engine(
  url = "https://neo4j.monarchinitiative.org",
  api_url = "https://api.monarchinitiative.org/v3/api"
)
}
\value{
An object of class \code{neo4j_engine} with the URL set to the Monarch Neo4j instance
}
\description{
Create a new neo4j engine object for the Monarch Neo4j instance
}
\examples{
g <- monarch_engine() \%>\%
 search_kg("Cystic fibrosis", limit = 1) \%>\%
 fetch_edges(predicate = "biolink:subclass_of")
g
}
