% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_search.R
\name{monarch_search}
\alias{monarch_search}
\title{Search the Monarch Initiative}
\usage{
monarch_search(query, category = NULL, limit = 10)
}
\arguments{
\item{query}{A character string representing the query term. Example terms:
"Cystic Fibrosis", "CYP6B", "swelling of joints".}

\item{category}{A character string indicating a single entity category in which to search for the query term.
Defaults to NULL, to search in any category.}

\item{limit}{An integer indicating the maximum number of search results to return. Defaults to 10.}
}
\value{
A \code{tbl_kgx} graph object containing the search results as nodes, with no edges.
}
\description{
This function uses the Monarch Initiative API to search for entities matching
a given query string within a specified category, returing a graph containg just the nodes in the search results.
}
\details{
The returned graph will contain only nodes with no associations between them, even if they exist in the Monarch database.
}
\examples{
monarch_search("Cystic Fibrosis", "biolink:Disease", 5)

}
