% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monarch_search.R
\name{monarch_search}
\alias{monarch_search}
\title{Search the Monarch Initiative}
\usage{
monarch_search(query, category = "biolink:Disease", limit = 10, offset = 0)
}
\arguments{
\item{query}{A character string representing the query term. Example terms:
"Cystic Fibrosis", "CYP6B", "swelling of joints".}

\item{category}{A character string indicating the entity category in which to search for the query term.
Defaults to "biolink:Disease".}

\item{limit}{An integer indicating the maximum number of search results to return. Defaults to 10.}

\item{offset}{An integer indicating the number of search results to skip (for pagination). Defaults to 0.}
}
\value{
A \code{monarch_kg} graph object containing the search results as nodes, with no edges.
}
\description{
This function uses the Monarch Initiative API to search for entities matching
a given query string within a specified category.
}
\details{
The returned graph will contain only nodes with no associations between them, even if they exist in the Monarch database.
}
\examples{
monarch_search("Cystic Fibrosis", "biolink:Disease", 5)

}
