% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_engine.R
\name{neo4j_engine}
\alias{neo4j_engine}
\title{neo4j_engine}
\usage{
neo4j_engine(url, username = NA, password = NA, preferences = NULL, ...)
}
\arguments{
\item{url}{A character string indicating the URL of the neo4j database.}

\item{username}{A character string indicating the username for the neo4j database (if needed).}

\item{password}{A character string indicating the password for the neo4j database (if needed).}

\item{preferences}{A named list of preferences for the engine.}

\item{...}{Additional arguments passed to neo2R::startGraph.}
}
\value{
An object of class \code{neo4j_engine}
}
\description{
Create a new neo4j engine object
}
\examples{
e <- neo4j_engine(url = "https://neo4j.monarchinitiative.org")
}
