% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes}
\alias{nodes}
\title{Get graph nodes table.}
\usage{
nodes(graph, ...)
}
\arguments{
\item{graph}{A graph object}

\item{...}{Other options (unused)}
}
\value{
A tibble with the nodes of the graph
}
\description{
Get graph nodes table.
}
\examples{
# (using the MONDO KGX file packaged with monarchr)
filename <- system.file("extdata", "mondo_kgx_tsv.tar.gz", package = "monarchr")

g <- file_engine(filename) |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

print(nodes(g))
}
