% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes}
\alias{nodes}
\title{Get graph nodes table.}
\usage{
nodes(graph, ...)
}
\arguments{
\item{graph}{A graph object}

\item{...}{Other options (unused)}
}
\value{
A tibble with the nodes of the graph
}
\description{
Get graph nodes table.
}
\examples{
# (using the example KGX file packaged with monarchr)
data(eds_marfan_kg)

g <- eds_marfan_kg |>
  fetch_nodes(query_ids = c("MONDO:0007525", "MONDO:0007526"))

print(nodes(g))
}
