% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_query.neo4j_engine.R
\name{normalize_categories}
\alias{normalize_categories}
\title{Normalize Categories}
\usage{
normalize_categories(cats_list, cats_prefs)
}
\arguments{
\item{cats_list}{A list of vectors of categories.}

\item{cats_prefs}{An ordered preference list over categories.}
}
\value{
A vector of normalized categories.
}
\description{
This function takes a list of vectors of categories and an ordered preference list over categories.
It selects the most preferred category from each vector, or the first category if no preferred categories are included.
}
\examples{
categories_list <- list(c("A", "B", "C"), c("D", "E", "F"))
categories_prefs <- c("B", "E", "A", "D", "C", "F")
normalize_categories(categories_list, categories_prefs)
}
