% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tbl_kgx.R
\name{plot.tbl_kgx}
\alias{plot.tbl_kgx}
\title{Specialized plot() function for KGX graphs}
\usage{
\method{plot}{tbl_kgx}(g, ...)
}
\arguments{
\item{g}{A tbl_kgx graph.}

\item{...}{other parameters (not used).}
}
\description{
Specialized plot() function for KGX graphs
}
