% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tbl_kgx.R
\name{plot.tbl_kgx}
\alias{plot.tbl_kgx}
\title{Specialized plot() function for KGX graphs}
\usage{
\method{plot}{tbl_kgx}(
  g,
  layout = "sugiyama",
  node_color = pcategory,
  edge_color = predicate,
  ...
)
}
\arguments{
\item{g}{A tbl_kgx graph.}

\item{layout}{The layout to use for the plot. Default is "sugiyama".}

\item{node_color}{The column to use for node color. Default is pcategory.}

\item{edge_color}{The column to use for edge color. Default is predicate.}

\item{...}{other parameters (not used).}
}
\description{
Specialized plot() function for KGX graphs
}
