% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tbl_kgx.R
\name{plot.tbl_kgx}
\alias{plot.tbl_kgx}
\title{Specialized \code{plot()} function for KGX graphs}
\usage{
\method{plot}{tbl_kgx}(
  g,
  ...,
  layout = "auto",
  node_color = pcategory,
  node_shape = namespace,
  edge_color = predicate,
  edge_linetype = primary_knowledge_source,
  plot_ids = FALSE,
  label_size = 2,
  fan_strength = 2,
  edge_alpha = 0.9,
  node_alpha = 0.9
)
}
\arguments{
\item{g}{A \link{tbl_kgx} graph.}

\item{...}{
  Arguments passed on to \code{\link[ggraph:ggraph]{ggraph::ggraph}}
  \describe{
    \item{\code{graph}}{The object containing the graph. See \emph{Details} for a list
of supported classes. Or a \code{layout_ggraph} object as returned from
\code{create_layout} in which case all subsequent arguments is ignored.}
  }}

\item{layout}{The layout to use for the plot. Default is "auto" as used by \code{ggraph}.}

\item{node_color}{The column to use for node color. Default is "pcategory".}

\item{node_shape}{The column to use for node shape Default is "namespace".}

\item{edge_color}{The column to use for edge color. Default is "predicate".}

\item{edge_linetype}{The column to use for edge line type. Default is "primary_knowledge_source".}

\item{plot_ids}{Whether to plot node IDs in addition to the default of "name". Defaults to FALSE.}

\item{label_size}{Size of node label text. Default is 2.}

\item{fan_strength}{Fan strength in ggraph's geom_edge_fan, Default is 2.}

\item{edge_alpha}{Alpha value for edges, default 0.9.}

\item{node_alpha}{Alpha value for nodes, default 0.9.}
}
\description{
Specialized \code{plot()} function for KGX graphs
}
\examples{
filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")
g <- file_engine(filename) |>
          fetch_nodes(query_ids = "MONDO:0007525") |>
          expand(predicates = "biolink:has_phenotype",
                 categories = "biolink:PhenotypicFeature")|>
          expand(categories = "biolink:Gene")
plot(g)
}
