% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_kg.R
\name{search_kg}
\alias{search_kg}
\title{search_kg}
\usage{
search_kg(obj, ...)
}
\arguments{
\item{obj}{a graph engine object}

\item{...}{additional parameters}
}
\value{
a tbl_kgx graph
}
\description{
Search a knowledge graph using a specific engine. Different engines
support different search capabilities and parameters.
}
\examples{
# the monarch engine uses the Monarch initiative search api
e <- monarch_engine()
g <- search_kg(e, "cystic fibrosis", category = "biolink:Disease", limit = 20)

# the neo4j engine searches a given property for regex matches
e <- neo4j_engine(url = "https://neo4j.monarchinitiative.org")
g <- search_kg(e, "MONDO:.*", field = "id", limit = 20)

}
