% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood.R
\name{summarize_neighborhood}
\alias{summarize_neighborhood}
\title{Summarize neighborhood}
\usage{
summarize_neighborhood(g, direction = "both", summarize = "edges")
}
\arguments{
\item{g}{A graph}

\item{direction}{The direction of edges to include in the neighborhood}

\item{summarize}{Whether to summarize edges, nodes, or both}
}
\value{
A tbl_kgx graph
}
\description{
Summarizes the neighborhood of the nodes of a given graph. Specifically,
letting $N$ be the set of nodes in the given graph, this function retrieves
counts of relationship predicates (or node categories) of nodes connected to
$N$ but not in $N$. This can be useful to examine the scale and scope of a
graph's collective neighborhood in the larger KG.
}
\details{
Note that the number of relationships returned may be larger than the number
of nodes they connect to; use summarize = "edges" to see edge counts between
nodes of different categories, and summarize = "nodes" to see counts of connected
node categories.

It is also possible to specify the direction of edges to include in the
neighborhood, using the direction parameter. The default is "both", which
includes both incoming and outgoing edges.
}
\examples{
g <- monarch_search("fanconi anemia")
summarize_neighborhood(g, summarize = "edges") # default
summarize_neighborhood(g, summarize = "nodes")

}
