% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood.R
\name{summarize_neighbornood}
\alias{summarize_neighbornood}
\alias{summarize_neighborhood}
\title{Summarize neighborhood}
\usage{
summarize_neighbornood(g, direction = "both", limit = 5)
}
\arguments{
\item{g}{A graph}

\item{direction}{The direction of edges to include in the neighborhood}

\item{limit}{The number of nodes to include in the neighborhood}
}
\value{
A tbl_kgx graph
}
\description{
Summarize the neighborhood of a set of nodes in a graph
}
\examples{
g <- monarch_search("fanconi anemia", limit = 5)
summarize_neighborhood(g)

}
\keyword{internal}
