% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood_nodes.R
\name{summarize_neighborhood_nodes}
\alias{summarize_neighborhood_nodes}
\title{Summarize neighborhood nodes}
\usage{
summarize_neighborhood_nodes(obj, ...)
}
\arguments{
\item{obj}{An object of class 'neighborhood'.}

\item{...}{Additional arguments to be passed to methods.}
}
\description{
This is a generic function to summarize neighborhood nodes. It is used by the
'summarize_neighborhood' function to summarize the neighborhood nodes of the
input object.
}
\examples{
g <- monarch_search("fanconi anemia")
summarize_neighborhood_nodes(g)

}
