% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_neighborhood_nodes.R
\name{summarize_neighborhood_nodes}
\alias{summarize_neighborhood_nodes}
\title{Summarize neighborhood nodes}
\usage{
summarize_neighborhood_nodes(obj, ...)
}
\arguments{
\item{e}{A graph engine object (optional - the recent engine used by the graph will be used if not supplied)}

\item{g}{A graph}

\item{direction}{The direction of edges to follow in the neighborhood}
}
\value{
A tbl_kgx graph
}
\description{
Summarizes the nodes in the neighborhood of a given graph.
}
\examples{
\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
g <- monarch_engine() |>
  search_nodes("fanconi anemia", limit = 5) |>
  summarize_neighborhood_nodes(direction = "both")
\dontshow{\}) # examplesIf}
}
