% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.neo4j_engine.R
\name{summary.neo4j_engine}
\alias{summary.neo4j_engine}
\title{Summarize contents of a Neo4j KG engine}
\usage{
\method{summary}{neo4j_engine}(engine, ..., quiet = FALSE)
}
\arguments{
\item{engine}{A \code{neo4j_engine} object}

\item{...}{Other parameters (not used)}

\item{quiet}{Logical, whether to suppress printing of the summary}
}
\value{
A list of dataframes
}
\description{
Given a Neo4j based KG engine, provides summary information in the form of
node counts, category counts across nodes, and relationship type counts.
General information about the graph is printed to the console, and a list of
dataframes describing node and edge counts is returned invisibly.
}
\examples{
\dontshow{if (monarch_engine_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# prints a readable summary and returns a list of dataframes
stats <- monarch_engine() |> summary()
print(stats)
\dontshow{\}) # examplesIf}
}
