% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_kgx.R
\name{tbl_kgx}
\alias{tbl_kgx}
\title{Create a KGX graph object}
\usage{
tbl_kgx(nodes = NULL, edges = NULL, ...)
}
\arguments{
\item{nodes}{A data frame containing the nodes of the graph. Must have 'id' and 'category' columns.}

\item{edges}{A data frame containing the edges of the graph. Must have 'subject', 'predicate', and 'object' columns. Can be NULL.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A KGX graph object.
}
\description{
This function creates a new tbl_kgx object which inherits from tidygraph::tbl_graph, from node and edge dataframes, ensuring they conform to the KGX specification
described at https://github.com/biolink/kgx/blob/master/specification/kgx-format.md. Specifically, nodes must have an 'id' and 'category' column,
and edges, if provided, must have 'subject', 'predicate', and 'object' columns. The function allows graphs with no edges.
The function sets 'from' and 'to' columns in the edges from 'subject' and 'object' respectively, and sets the node key to 'id'.
Additional columns are allowed, and a logical \code{selected} column is added to the nodes data frame to indicate whether a node is selected, if not already present.
If a selected column is present, it should be a logical vector. If not, an error is thrown.
}
\details{
This function will generally be called internally.
}
\examples{
nodes <- data.frame(id = c("A", "B"), category = c("gene", "disease"))
edges <- data.frame(subject = c("A"), predicate = c("associated_with"), object = c("B"))
g <- tbl_kgx(nodes, edges)
}
