% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitive_closure.R
\name{transitive_closure}
\alias{transitive_closure}
\title{Compute transitive closure over a predicate.}
\usage{
transitive_closure(g, predicate = "biolink:subclass_of")
}
\arguments{
\item{g}{The \code{tbl_kgx} graph to compute on.}

\item{predicate}{The edge predicate to close over.}
}
\value{
Graph with transitive edges added.
}
\description{
Computes the transitive closure of a graph, treating the specified
predicate as transitive. Resulting edge predicates will the be the
same, but have primary_knowledge_source set to transitive_\if{html}{\out{<predicate>}}.
}
\examples{
data(eds_marfan_kg)

eds_marfan_kg |> fetch_nodes(name == "Tall stature") |>
  # get 2 levels of ancestors
  expand_n(predicates = "biolink:subclass_of", direction = "out", n = 3) |>
  activate(edges) |>
  filter(primary_knowledge_source == "infores:upheno") |>
  transitive_closure(predicate = "biolink:subclass_of") |>
  plot(edge_color = primary_knowledge_source)

}
\seealso{
\code{\link[=roll_up]{roll_up()}}, \code{\link[=transfer]{transfer()}}, \code{\link[=descendants]{descendants()}}, \code{\link[=ancestors]{ancestors()}}
}
